+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [UNRELEASED]
 
-No user facing changes.
+- Update default CodeQL bundle version to 2.11.4. [#1382](https://github.com/github/codeql-action/pull/1382)
 
 ## 2.1.33 - 16 Nov 2022